requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
flatbuffers==24.12.23
fonttools==4.55.3
future==0.18.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
idna==3.10
imageio==2.37.0
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-cupti-cu12==12.5.82
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.5.82
nvidia-cuda-runtime-cu12==12.5.82
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu12==11.2.3.61
nvidia-curand-cu12==10.3.6.82
nvidia-cusolver-cu12==11.6.3.83
nvidia-cusparse-cu12==12.5.1.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.5.82
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
Pygments==2.19.1
pyimage==0.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.9.4
scikit-image==0.25.0
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
tifffile==2025.1.10
tk==0.1.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
